Centralize UI tests to UIFixture
