+++ src/wix/test/WixToolsetTest.CoreIntegration/MsiQueryFixture.cs
@@ -290,58 +290,6 @@ namespace WixToolsetTest.CoreIntegration
                 }, results);
             }
         }
-
-        [Fact]
-        public void PopulatesControlTables()
-        {
-            var folder = TestData.Get(@"TestData");
-
-            using (var fs = new DisposableFileSystem())
-            {
-                var baseFolder = fs.GetFolder();
-                var intermediateFolder = Path.Combine(baseFolder, "obj");
-                var msiPath = Path.Combine(baseFolder, @"bin\test.msi");
-
-                var result = WixRunner.Execute(new[]
-                {
-                    "build",
-                    Path.Combine(folder, "DialogsInInstallUISequence", "PackageComponents.wxs"),
-                    Path.Combine(folder, "ProductWithComponentGroupRef", "MinimalComponentGroup.wxs"),
-                    Path.Combine(folder, "ProductWithComponentGroupRef", "Product.wxs"),
-                    "-bindpath", Path.Combine(folder, "SingleFile", "data"),
-                    "-intermediateFolder", intermediateFolder,
-                    "-o", msiPath,
-                });
-
-                result.AssertSuccess();
-
-                Assert.True(File.Exists(msiPath));
-
-                var results = Query.QueryDatabase(msiPath, new[] { "CheckBox", "Control", "ControlCondition", "InstallUISequence" });
-                WixAssert.CompareLineByLine(new[]
-                {
-                    "CheckBox:WIXUI_EXITDIALOGOPTIONALCHECKBOX\t1",
-                    "Control:FirstDialog\tHeader\tText\t0\t13\t90\t13\t3\t\tFirstDialogHeader\tTitle\t",
-                    "Control:FirstDialog\tTitle\tText\t0\t0\t90\t13\t3\t\tFirstDialogTitle\tHeader\t",
-                    "Control:SecondDialog\tOptionalCheckBox\tCheckBox\t0\t13\t100\t40\t2\tWIXUI_EXITDIALOGOPTIONALCHECKBOX\t[WIXUI_EXITDIALOGOPTIONALCHECKBOXTEXT]\tTitle\tOptional checkbox|Check this box for fun",
-                    "Control:SecondDialog\tTitle\tText\t0\t0\t90\t13\t3\t\tSecondDialogTitle\tOptionalCheckBox\t",
-                    "ControlCondition:FirstDialog\tHeader\tDisable\tInstalled",
-                    "ControlCondition:FirstDialog\tHeader\tHide\tInstalled",
-                    "ControlCondition:SecondDialog\tOptionalCheckBox\tShow\tWIXUI_EXITDIALOGOPTIONALCHECKBOXTEXT AND NOT Installed",
-                    "InstallUISequence:CostFinalize\t\t1000",
-                    "InstallUISequence:CostInitialize\t\t800",
-                    "InstallUISequence:ExecuteAction\t\t1300",
-                    "InstallUISequence:FileCost\t\t900",
-                    "InstallUISequence:FindRelatedProducts\t\t25",
-                    "InstallUISequence:FirstDialog\tInstalled AND PATCH\t1298",
-                    "InstallUISequence:LaunchConditions\t\t100",
-                    "InstallUISequence:MigrateFeatureStates\t\t1200",
-                    "InstallUISequence:SecondDialog\tNOT Installed\t1299",
-                    "InstallUISequence:ValidateProductID\t\t700",
-                }, results);
-            }
-        }
-
         [Fact]
         public void PopulatesCreateFolderTableForNullKeypathComponents()
         {
